Support multi bookies
