+++ herddb-core/src/test/java/herddb/cluster/BackupRestoreTest.java
@@ -72,7 +72,7 @@ public class BackupRestoreTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
