+++ herddb-core/src/test/java/herddb/cluster/MultiServerCreateTableSpaceWaitTest.java
@@ -53,7 +53,7 @@ public class MultiServerCreateTableSpaceWaitTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
