+++ herddb-core/src/test/java/herddb/cluster/RetryOnLeaderChangedTest.java
@@ -66,7 +66,7 @@ public class RetryOnLeaderChangedTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
