+++ herddb-core/src/test/java/herddb/cluster/TablespaceReplicasStateTest.java
@@ -60,7 +60,7 @@ public class TablespaceReplicasStateTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
