+++ herddb-core/src/test/java/herddb/cluster/ZookeeperMetadataStorageManagerTest.java
@@ -51,7 +51,7 @@ public class ZookeeperMetadataStorageManagerTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
