+++ herddb-core/src/test/java/herddb/cluster/bookkeeper/BookkeeperFailuresBase.java
@@ -44,7 +44,7 @@ public abstract class BookkeeperFailuresBase {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
