+++ herddb-core/src/test/java/herddb/cluster/follower/MultiServerBase.java
@@ -42,7 +42,7 @@ public class MultiServerBase {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
