+++ herddb-core/src/test/java/herddb/server/DownloadTableSpaceTest.java
@@ -67,7 +67,7 @@ public class DownloadTableSpaceTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
