+++ herddb-core/src/test/java/herddb/server/LedgerManagementTest.java
@@ -55,7 +55,7 @@ public class LedgerManagementTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
