+++ herddb-core/src/test/java/herddb/server/MaxLeaderInactivityTest.java
@@ -52,7 +52,7 @@ public class MaxLeaderInactivityTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
