+++ herddb-core/src/test/java/herddb/server/UseVirtualTableSpaceIdWithZookKeeperTest.java
@@ -66,7 +66,7 @@ public class UseVirtualTableSpaceIdWithZookKeeperTest {
     @Before
     public void beforeSetup() throws Exception {
         testEnv = new ZKTestEnv(folder.newFolder().toPath());
-        testEnv.startBookie();
+        testEnv.startBookieAndInitCluster();
     }
 
     @After
