+++ herddb-core/src/test/java/herddb/utils/ZKTestEnv.java
@@ -42,7 +42,7 @@ public class ZKTestEnv implements AutoCloseable {
     List<BookieServer> bookies = new ArrayList<>();
     Path path;
     private int nextBookiePort = 5621;
-    
+
 
     public ZKTestEnv(Path path) throws Exception {
         zkServer = new TestingServer(1282, path.toFile(), true);
@@ -50,7 +50,7 @@ public class ZKTestEnv implements AutoCloseable {
 
         try (ZooKeeperClient zkc = ZooKeeperClient
                 .newBuilder()
-                .connectString("localhost:1282")
+                .connectString(zkServer.getConnectString())
                 .sessionTimeoutMs(10000)
                 .build()) {
 
@@ -62,39 +62,19 @@ public class ZKTestEnv implements AutoCloseable {
         }
     }
 
-    public void startBookie() throws Exception {
+    public void startBookieAndInitCluster() throws Exception {
         startBookie(true);
     }
-
-    public void startBookie(boolean format) throws Exception {        
-        ServerConfiguration conf = new ServerConfiguration();
-        conf.setBookiePort(nextBookiePort++);
-        System.out.println("STARTING BOOKIE at port " + nextBookiePort);
-        conf.setUseHostNameAsBookieID(true);
-
-        // no need to preallocate journal and entrylog in tests
-        conf.setEntryLogFilePreAllocationEnabled(false);
-        conf.setProperty("journalPreAllocSizeMB", 1);
-
-        Path targetDir = path.resolve("bookie_data_"+conf.getBookiePort());
-        conf.setMetadataServiceUri("zk+null://" + zkServer.getConnectString() + herddb.server.ServerConfiguration.PROPERTY_BOOKKEEPER_LEDGERS_PATH_DEFAULT);
-        conf.setLedgerDirNames(new String[]{targetDir.toAbsolutePath().toString()});
-        conf.setJournalDirName(targetDir.toAbsolutePath().toString());
-        conf.setFlushInterval(10000);
-        conf.setGcWaitTime(5);
-        conf.setJournalFlushWhenQueueEmpty(true);
-//        conf.setJournalBufferedEntriesThreshold(1);
-        conf.setAutoRecoveryDaemonEnabled(false);
-
-        // no need for real network in tests
-        conf.setEnableLocalTransport(true);
-        conf.setDisableServerSocketBind(true);
-
-        // no need to fsync in tests
-        conf.setJournalSyncData(false);
-
-        conf.setAllowLoopback(true);
-        conf.setProperty("journalMaxGroupWaitMSec", 10); // default 200ms
+    
+    public void startNewBookie() throws Exception {
+        startBookie(false);
+    }
+    
+    private void startBookie(boolean format) throws Exception {
+        if (format && !bookies.isEmpty()) {
+            throw new Exception("cannot format, you aleady have bookies");
+        }
+        ServerConfiguration conf = createBookieConf(nextBookiePort++);
 
         if (format) {
             BookKeeperAdmin.initNewCluster(conf);
@@ -105,12 +85,57 @@ public class ZKTestEnv implements AutoCloseable {
         bookies.add(bookie);
         bookie.start();
     }
-    
+
+    private ServerConfiguration createBookieConf(int port) {
+        ServerConfiguration conf = new ServerConfiguration();
+        conf.setBookiePort(port++);
+        System.out.println("STARTING BOOKIE at port " + port);
+        conf.setUseHostNameAsBookieID(true);
+        // no need to preallocate journal and entrylog in tests
+        conf.setEntryLogFilePreAllocationEnabled(false);
+        conf.setProperty("journalPreAllocSizeMB", 1);
+        Path targetDir = path.resolve("bookie_data_"+conf.getBookiePort());
+        conf.setMetadataServiceUri("zk+null://" + zkServer.getConnectString() + herddb.server.ServerConfiguration.PROPERTY_BOOKKEEPER_LEDGERS_PATH_DEFAULT);
+        conf.setLedgerDirNames(new String[]{targetDir.toAbsolutePath().toString()});
+        conf.setJournalDirName(targetDir.toAbsolutePath().toString());
+        conf.setFlushInterval(10000);
+        conf.setGcWaitTime(5);
+        conf.setJournalFlushWhenQueueEmpty(true);
+        //        conf.setJournalBufferedEntriesThreshold(1);
+        conf.setAutoRecoveryDaemonEnabled(false);
+        // no need for real network in tests
+        conf.setEnableLocalTransport(true);
+        conf.setDisableServerSocketBind(true);
+        // no need to fsync in tests
+        conf.setJournalSyncData(false);
+        conf.setAllowLoopback(true);
+        conf.setProperty("journalMaxGroupWaitMSec", 10); // default 200ms
+        return conf;
+    }
+
+    public void startStoppedBookie(String addr) throws Exception {
+        int index = 0;
+        for (BookieServer bookie : bookies) {
+            if (bookie.getLocalAddress().getSocketAddress().toString().equals(addr)) {
+                if (bookie.isRunning()) {
+                    throw new Exception("you did not stop bookie "+addr);
+                }
+                ServerConfiguration conf = createBookieConf(bookie.getLocalAddress().getPort());
+                BookieServer newBookie = new BookieServer(conf);
+                bookies.set(index, newBookie);
+                newBookie.start();
+                return;
+            }
+            index++;
+        }
+        throw new Exception("Cannot find bookie "+addr);
+    }
+
     public void pauseBookie() throws Exception {
         bookies.get(0).suspendProcessing();
     }
-    
-    public void pauseBookie(String addr) throws Exception {        
+
+    public void pauseBookie(String addr) throws Exception {
         for (BookieServer bookie : bookies) {
             if (bookie.getLocalAddress().getSocketAddress().toString().equals(addr)) {
                 bookie.suspendProcessing();
@@ -123,8 +148,8 @@ public class ZKTestEnv implements AutoCloseable {
     public void resumeBookie() throws Exception {
         bookies.get(0).resumeProcessing();
     }
-    
-    public void resumeBookie(String addr) throws Exception {        
+
+    public void resumeBookie(String addr) throws Exception {
         for (BookieServer bookie : bookies) {
             if (bookie.getLocalAddress().getSocketAddress().toString().equals(addr)) {
                 bookie.resumeProcessing();
@@ -132,18 +157,27 @@ public class ZKTestEnv implements AutoCloseable {
             }
         }
         throw new Exception("Cannot find bookie "+addr);
-    }        
-
-    public void stopBookie() throws Exception {
-        for (BookieServer bookie : bookies) {
-            bookie.shutdown();
-            bookie.join();         
-        }
-        bookies.clear();
     }
 
+    public String stopBookie() throws Exception {
+        String addr = bookies.get(0).getLocalAddress().getSocketAddress().toString();
+        stopBookie(addr);
+        return addr;
+    }
+    
+    public void stopBookie(String addr) throws Exception {
+        for (BookieServer bookie : bookies) {
+            if (bookie.getLocalAddress().getSocketAddress().toString().equals(addr)) {
+                bookie.shutdown();
+                bookie.join();
+                return;
+            }
+        }
+        throw new Exception("Cannot find bookie "+addr);
+    }    
+
     public String getAddress() {
-        return "localhost:1282";
+        return zkServer.getConnectString();
     }
 
     public int getTimeout() {
@@ -162,10 +196,12 @@ public class ZKTestEnv implements AutoCloseable {
             } catch (Throwable t) {
             }
         }
-        
+        bookies.clear();
+
         try {
             if (zkServer != null) {
                 zkServer.close();
+                zkServer = null;
             }
         } catch (Throwable t) {
         }
